pyproject.toml
[project]
name = "backend-python-ebac"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor Moser",email = "igormoser@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "redis (>=7.2.1,<8.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "kafka-python (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
